requirements.txt
python-dotenv
wolframalpha
aiohttp
langchain-google-genai
langchain-core
langchain-community
fastmcp
llama-cpp-python
google-generativeai
langchain